panel-launchers: Rework for the new cinnamon menus
To make sure alignment of the menu is correct with the launcher make sure they
are fully expanded to width/height of the panel.

Also change all the style class names since we need a bit different control over
these for the new layouts that are allowed with vertical panels. Changing to
new style class names allows us to mark these as important so we can keep them
working properly with themes that aren't up to date with Cinnamon3.2