popupMenu: Fix invisible main menu
Make sure all the animated values are set to their default when we are not
running an animation in the the open() function. Otherwise in some cases when
the main menu is hidden it's scale and opacity are left set to 0 and you cannot
see it next the next time it gets opened.