popupMenu: Ensure menus have a background color
Since we removed the boxpointer which sets the background color for menus in
most themes, many are left with a transparent menu. Add a backup code path to
set a color for menus in themes that haven't adapted to these changes yet. This
shouldn't be a permanent solution, but a temporary workaround for a version or
two to give themes time to adapt.