st-bin: Use child's align properties
This is the first step to clean up the mess of trying to create layouts
with St. StBin is one of the worst offenders since it mirrors Clutter
Actors own x-align and y-align properties. But these behave completely
different and even use a different enum for their values.

This will likely cause a small bit of fallout but should not prevent
anything from actually working.